# Conflicts:
#	src/main/java/com/levelup/levelup_academy/Config/ConfigurationSecurity.java
#	src/main/java/com/levelup/levelup_academy/Controller/ContractController.java
#	src/main/java/com/levelup/levelup_academy/DTO/ModeratorDTO.java
#	src/main/java/com/levelup/levelup_academy/Service/ContractService.java

+++ src/main/java/com/levelup/levelup_academy/Service/ProService.java
@@ -3,13 +3,10 @@ package com.levelup.levelup_academy.Service;
 import com.levelup.levelup_academy.Api.ApiException;
 import com.levelup.levelup_academy.DTO.EmailRequest;
 import com.levelup.levelup_academy.DTO.ProDTO;
-import com.levelup.levelup_academy.Model.Moderator;
-import com.levelup.levelup_academy.Model.Pro;
-import com.levelup.levelup_academy.Model.User;
-import com.levelup.levelup_academy.Repository.AuthRepository;
-import com.levelup.levelup_academy.Repository.ModeratorRepository;
-import com.levelup.levelup_academy.Repository.ProRepository;
+import com.levelup.levelup_academy.Model.*;
+import com.levelup.levelup_academy.Repository.*;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -28,15 +25,34 @@ public class ProService {
     private final AuthRepository authRepository;
     private final EmailNotificationService emailNotificationService;
     private final ModeratorRepository moderatorRepository;
+    private final SessionRepository sessionRepository;
+    private final ContractRepository contractRepository;
 
-    //GET
-    public List<Pro> getAllPro(){
+    //GET All Pro players by moderator
+    public List<Pro> getAllPro(Integer moderatorId) {
+        Moderator moderator= moderatorRepository.findModeratorById(moderatorId);
+        if(moderator == null){
+            throw new ApiException("Moderator not found");
+        }
         return proRepository.findAll();
     }
 
+    public Pro getPro(Integer moderatorId,Integer proId){
+        Moderator moderator= moderatorRepository.findModeratorById(moderatorId);
+        if(moderator == null){
+            throw new ApiException("Moderator not found");
+        }
+        Pro pro = proRepository.findProById(proId);
+        if(pro == null){
+            throw new ApiException("Pro player is not found");
+        }
+        return pro;
+    }
+
+
     //Register pro player
 
-    public void registerPro(ProDTO proDTO, MultipartFile file){
+    public void registerPro(ProDTO proDTO, MultipartFile file) {
         proDTO.setRole("PRO");
         String filePath = null;
         if (file != null && !file.isEmpty()) {
@@ -50,16 +66,17 @@ public class ProService {
                 throw new RuntimeException("Failed to save CV file.");
             }
         }
-        User user = new User(null, proDTO.getUsername(), proDTO.getPassword(), proDTO.getEmail(), proDTO.getFirstName(), proDTO.getLastName(), proDTO.getRole(), LocalDate.now(),null,null,null,null,null,null,null,null);
+        String hashPassword = new BCryptPasswordEncoder().encode(proDTO.getPassword());
+        User user = new User(null, proDTO.getUsername(), hashPassword, proDTO.getEmail(), proDTO.getFirstName(), proDTO.getLastName(), proDTO.getRole(), LocalDate.now(),null,null,null,null,null,null,null,null);
         Pro pro = new Pro(null, filePath, user, null, null,null,false);
         authRepository.save(user);
         proRepository.save(pro);
     }
 
 
-    public void edit(Integer proId,ProDTO proDTO){
+    public void edit(Integer proId, ProDTO proDTO) {
         Pro pro = proRepository.findProById(proId);
-        if(pro == null){
+        if (pro == null) {
             throw new ApiException("The Professional Player you search for is not found ");
         }
 
@@ -74,6 +91,8 @@ public class ProService {
         if (authRepository.existsByUsername(proDTO.getUsername()) && !pro.getUser().getUsername().equals(proDTO.getUsername())) {
             throw new ApiException("Username is already in use");
         }
+        String hashPassword = new BCryptPasswordEncoder().encode(proDTO.getPassword());
+        pro.getUser().setPassword(hashPassword);
         pro.getUser().setEmail(proDTO.getEmail());
         pro.getUser().setUsername(proDTO.getUsername());
 
@@ -88,13 +107,19 @@ public class ProService {
             throw new ApiException("The player is not found");
         }
         User user = pro.getUser();
-
+        if (pro.getIsApproved().equals(false)) {
+            throw new ApiException("The Professional Player is not approved yet ");
+        }
         proRepository.delete(pro);
         authRepository.delete(user);
     }
 
 
-    public byte[] downloadProCv(Integer proId) {
+    public byte[] downloadProPDF(Integer moderateId,Integer proId) {
+        Moderator moderator = moderatorRepository.findModeratorById(moderateId);
+        if(moderator == null){
+            throw new ApiException("Moderator not found");
+        }
         Pro pro = proRepository.findById(proId)
                 .orElseThrow(() -> new RuntimeException("Pro not found"));
 
@@ -129,7 +154,6 @@ public class ProService {
             throw new ApiException("This player has already been approved.");
         }
         User user = pro.getUser();
-//        user.setIsApproved(true);
         pro.setIsApproved(true);
         authRepository.save(user);
         proRepository.save(pro);
@@ -147,7 +171,6 @@ public class ProService {
         }
 
         User user = pro.getUser();
-//        user.setIsApproved(false);
         pro.setIsApproved(false);
         proRepository.delete(pro);
         authRepository.delete(user);
@@ -155,6 +178,10 @@ public class ProService {
 
 //moderator can see all the requests
     public List<Pro> getAllProRequests(Integer moderatorId){
+        Moderator moderator = moderatorRepository.findModeratorById(moderatorId);
+        if(moderator == null){
+            throw new ApiException("Moderator not found");
+        }
         return proRepository.findByIsApproved(false);
     }
 
@@ -187,5 +214,31 @@ public class ProService {
 
 
 
+    public void expireAccount() {
+        LocalDate oneYearAgo = LocalDate.now().minusYears(1);
+
+        List<Contract> contracts = contractRepository.findByStartDateBefore(oneYearAgo);
+
+        for (Contract contract : contracts) {
+            if (contract.getStartDate().isBefore(oneYearAgo) && contract.getContractStatus().equals("Accepted")) {
+                Pro pro = contract.getPro();
+                sendAccountDeletionEmail(pro);
+
+                contractRepository.save(contract);  // Save the updated contract
+            }
+        }
+    }
+    // Method to send email to pro player about account deleting the account
+    private void sendAccountDeletionEmail(Pro pro) {
+        String subject = "Your Account Has Been Deleted Due to Contract Expiry";
+        String body = String.format("Dear %s,\n\nYour contract has expired, and your account has been deleted as a result. If you have any questions, please contact support.\n\nBest regards,\nTeam",pro.getUser().getEmail());
+
+        EmailRequest emailRequest = new EmailRequest();
+        emailRequest.setRecipient(pro.getUser().getEmail());
+        emailRequest.setSubject(subject);
+        emailRequest.setMessage(body);
+
+        emailNotificationService.sendEmail(emailRequest);
+    }
 
 }
